# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API para a plataforma SkillBridge — liga estudantes e projetos.",
        "title": "SkillBridge API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/users/me": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Devolve o perfil do utilizador autenticado",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Obter perfil próprio",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Atualiza os dados do perfil do utilizador autenticado",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Atualizar perfil",
                "parameters": [
                    {
                        "description": "Dados a atualizar",
                        "name": "input",
                        "in": "body",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "avatar_url": {
                                    "type": "string"
                                },
                                "bio": {
                                    "type": "string"
                                },
                                "course": {
                                    "type": "string"
                                },
                                "name": {
                                    "type": "string"
                                },
                                "university": {
                                    "type": "string"
                                },
                                "year": {
                                    "type": "integer"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/users/password-reset": {
            "post": {
                "description": "Gera um link Firebase e envia por email ao utilizador",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Pedir redefinição de palavra-passe",
                "parameters": [
                    {
                        "description": "Email do utilizador",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "email": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/users/register": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Cria um perfil de utilizador após autenticação Firebase",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Criar utilizador",
                "parameters": [
                    {
                        "description": "Dados do utilizador (email é lido automaticamente do token Firebase)",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "bio": {
                                    "type": "string"
                                },
                                "course": {
                                    "type": "string"
                                },
                                "name": {
                                    "type": "string"
                                },
                                "university": {
                                    "type": "string"
                                },
                                "year": {
                                    "type": "integer"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/users/{id}": {
            "get": {
                "description": "Devolve o perfil público de um utilizador",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Obter utilizador por ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID do utilizador",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.ContactLinks": {
            "type": "object",
            "properties": {
                "facebook": {
                    "type": "string"
                },
                "github": {
                    "type": "string"
                },
                "instagram": {
                    "type": "string"
                },
                "linkedin": {
                    "type": "string"
                },
                "tiktok": {
                    "type": "string"
                },
                "twitter": {
                    "type": "string"
                },
                "website": {
                    "type": "string"
                }
            }
        },
...
        "models.Project": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "members": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ProjectMember"
                    }
                },
                "owner": {
                    "$ref": "#/definitions/models.User"
                },
                "owner_id": {
                    "type": "integer"
                },
                "owners": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ProjectOwner"
                    }
                },
                "roles": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ProjectRole"
                    }
                },
                "status": {
                    "description": "open / in_progress / completed",
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "models.ProjectMember": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "joined_at": {
                    "type": "string"
                },
                "project_id": {
                    "type": "integer"
                },
                "role_id": {
                    "type": "integer"
                },
                "user": {
                    "$ref": "#/definitions/models.User"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "models.ProjectOwner": {
            "type": "object",
            "properties": {
                "added_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "project_id": {
                    "type": "integer"
                },
                "user": {
                    "$ref": "#/definitions/models.User"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "models.ProjectRole": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "filled": {
                    "type": "boolean"
                },
                "id": {
                    "type": "integer"
                },
                "project_id": {
                    "type": "integer"
                },
                "skill": {
                    "$ref": "#/definitions/models.Skill"
                },
                "skill_id": {
                    "type": "integer"
                }
            }
        },
        "models.Skill": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "models.User": {
            "type": "object",
            "properties": {
                "avatar_url": {
                    "type": "string"
                },
                "bio": {
                    "type": "string"
                },
                "contact_links": {
                    "$ref": "#/definitions/models.ContactLinks"
                },
                "course": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "firebase_uid": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "owned_projects": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Project"
                    }
                },
                "project_members": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ProjectMember"
                    }
                },
                "skills": {
                    "description": "Relações",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.UserSkill"
                    }
                },
                "university": {
                    "type": "string"
                },
                "year": {
                    "type": "integer"
                }
            }
        },
...
        "models.UserSkill": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "level": {
                    "description": "beginner / intermediate / advanced",
                    "type": "string"
                },
                "skill": {
                    "$ref": "#/definitions/models.Skill"
                },
                "skill_id": {
                    "type": "integer"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}